Refactor: Extract booking price recalculation to service
